working on some api stuff, fixed product findall to sort by id
